+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Foreign key preview: press Cmd+Enter on a FK cell to see the referenced row in a popover
+
 ## [0.27.2] - 2026-04-02
 
 ### Added